v5.8.1

+++ platform/Device/SiliconLabs/EFM32TG11B/Include/efm32tg11b540f64gm64.h
@@ -2,10 +2,10 @@
  * @file
  * @brief CMSIS Cortex-M Peripheral Access Layer Header File
  *        for EFM32TG11B540F64GM64
- * @version 5.7.3
+ * @version 5.8.1
  *******************************************************************************
  * # License
- * <b>Copyright 2018 Silicon Laboratories Inc. www.silabs.com</b>
+ * <b>Copyright 2019 Silicon Laboratories Inc. www.silabs.com</b>
  *******************************************************************************
  *
  * SPDX-License-Identifier: Zlib
@@ -137,64 +137,64 @@ typedef enum IRQn{
 #define PART_NUMBER                "EFM32TG11B540F64GM64" /**< Part Number */
 
 /** Memory Base addresses and limits */
-#define FLASH_MEM_BASE             ((uint32_t) 0x00000000UL) /**< FLASH base address  */
-#define FLASH_MEM_SIZE             ((uint32_t) 0x10000000UL) /**< FLASH available address space  */
-#define FLASH_MEM_END              ((uint32_t) 0x0FFFFFFFUL) /**< FLASH end address  */
-#define FLASH_MEM_BITS             ((uint32_t) 0x0000001CUL) /**< FLASH used bits  */
-#define RAM0_CODE_MEM_BASE         ((uint32_t) 0x10000000UL) /**< RAM0_CODE base address  */
-#define RAM0_CODE_MEM_SIZE         ((uint32_t) 0x8000UL)     /**< RAM0_CODE available address space  */
-#define RAM0_CODE_MEM_END          ((uint32_t) 0x10007FFFUL) /**< RAM0_CODE end address  */
-#define RAM0_CODE_MEM_BITS         ((uint32_t) 0x0000000FUL) /**< RAM0_CODE used bits  */
-#define CRYPTO0_MEM_BASE           ((uint32_t) 0x400F0000UL) /**< CRYPTO0 base address  */
-#define CRYPTO0_MEM_SIZE           ((uint32_t) 0x400UL)      /**< CRYPTO0 available address space  */
-#define CRYPTO0_MEM_END            ((uint32_t) 0x400F03FFUL) /**< CRYPTO0 end address  */
-#define CRYPTO0_MEM_BITS           ((uint32_t) 0x0000000AUL) /**< CRYPTO0 used bits  */
-#define PER_BITCLR_MEM_BASE        ((uint32_t) 0x44000000UL) /**< PER_BITCLR base address  */
-#define PER_BITCLR_MEM_SIZE        ((uint32_t) 0x50000UL)    /**< PER_BITCLR available address space  */
-#define PER_BITCLR_MEM_END         ((uint32_t) 0x4404FFFFUL) /**< PER_BITCLR end address  */
-#define PER_BITCLR_MEM_BITS        ((uint32_t) 0x00000013UL) /**< PER_BITCLR used bits  */
-#define PER1_BITCLR_MEM_BASE       ((uint32_t) 0x44050000UL) /**< PER1_BITCLR base address  */
-#define PER1_BITCLR_MEM_SIZE       ((uint32_t) 0xA0000UL)    /**< PER1_BITCLR available address space  */
-#define PER1_BITCLR_MEM_END        ((uint32_t) 0x440EFFFFUL) /**< PER1_BITCLR end address  */
-#define PER1_BITCLR_MEM_BITS       ((uint32_t) 0x00000014UL) /**< PER1_BITCLR used bits  */
-#define PER1_BITSET_MEM_BASE       ((uint32_t) 0x46050000UL) /**< PER1_BITSET base address  */
-#define PER1_BITSET_MEM_SIZE       ((uint32_t) 0xA0000UL)    /**< PER1_BITSET available address space  */
-#define PER1_BITSET_MEM_END        ((uint32_t) 0x460EFFFFUL) /**< PER1_BITSET end address  */
-#define PER1_BITSET_MEM_BITS       ((uint32_t) 0x00000014UL) /**< PER1_BITSET used bits  */
-#define CRYPTO0_BITSET_MEM_BASE    ((uint32_t) 0x460F0000UL) /**< CRYPTO0_BITSET base address  */
-#define CRYPTO0_BITSET_MEM_SIZE    ((uint32_t) 0x400UL)      /**< CRYPTO0_BITSET available address space  */
-#define CRYPTO0_BITSET_MEM_END     ((uint32_t) 0x460F03FFUL) /**< CRYPTO0_BITSET end address  */
-#define CRYPTO0_BITSET_MEM_BITS    ((uint32_t) 0x0000000AUL) /**< CRYPTO0_BITSET used bits  */
-#define CRYPTO0_BITCLR_MEM_BASE    ((uint32_t) 0x440F0000UL) /**< CRYPTO0_BITCLR base address  */
-#define CRYPTO0_BITCLR_MEM_SIZE    ((uint32_t) 0x400UL)      /**< CRYPTO0_BITCLR available address space  */
-#define CRYPTO0_BITCLR_MEM_END     ((uint32_t) 0x440F03FFUL) /**< CRYPTO0_BITCLR end address  */
-#define CRYPTO0_BITCLR_MEM_BITS    ((uint32_t) 0x0000000AUL) /**< CRYPTO0_BITCLR used bits  */
-#define PER_BITSET_MEM_BASE        ((uint32_t) 0x46000000UL) /**< PER_BITSET base address  */
-#define PER_BITSET_MEM_SIZE        ((uint32_t) 0x50000UL)    /**< PER_BITSET available address space  */
-#define PER_BITSET_MEM_END         ((uint32_t) 0x4604FFFFUL) /**< PER_BITSET end address  */
-#define PER_BITSET_MEM_BITS        ((uint32_t) 0x00000013UL) /**< PER_BITSET used bits  */
-#define PER_MEM_BASE               ((uint32_t) 0x40000000UL) /**< PER base address  */
-#define PER_MEM_SIZE               ((uint32_t) 0x50000UL)    /**< PER available address space  */
-#define PER_MEM_END                ((uint32_t) 0x4004FFFFUL) /**< PER end address  */
-#define PER_MEM_BITS               ((uint32_t) 0x00000013UL) /**< PER used bits  */
-#define PER1_MEM_BASE              ((uint32_t) 0x40050000UL) /**< PER1 base address  */
-#define PER1_MEM_SIZE              ((uint32_t) 0xA0000UL)    /**< PER1 available address space  */
-#define PER1_MEM_END               ((uint32_t) 0x400EFFFFUL) /**< PER1 end address  */
-#define PER1_MEM_BITS              ((uint32_t) 0x00000014UL) /**< PER1 used bits  */
-#define DEVICE_MEM_BASE            ((uint32_t) 0xF0040000UL) /**< DEVICE base address  */
-#define DEVICE_MEM_SIZE            ((uint32_t) 0x1000UL)     /**< DEVICE available address space  */
-#define DEVICE_MEM_END             ((uint32_t) 0xF0040FFFUL) /**< DEVICE end address  */
-#define DEVICE_MEM_BITS            ((uint32_t) 0x0000000CUL) /**< DEVICE used bits  */
-#define RAM0_MEM_BASE              ((uint32_t) 0x20000000UL) /**< RAM0 base address  */
-#define RAM0_MEM_SIZE              ((uint32_t) 0x8000UL)     /**< RAM0 available address space  */
-#define RAM0_MEM_END               ((uint32_t) 0x20007FFFUL) /**< RAM0 end address  */
-#define RAM0_MEM_BITS              ((uint32_t) 0x0000000FUL) /**< RAM0 used bits  */
+#define FLASH_MEM_BASE             (0x00000000UL) /**< FLASH base address  */
+#define FLASH_MEM_SIZE             (0x10000000UL) /**< FLASH available address space  */
+#define FLASH_MEM_END              (0x0FFFFFFFUL) /**< FLASH end address  */
+#define FLASH_MEM_BITS             (0x0000001CUL) /**< FLASH used bits  */
+#define RAM0_CODE_MEM_BASE         (0x10000000UL) /**< RAM0_CODE base address  */
+#define RAM0_CODE_MEM_SIZE         (0x8000UL)     /**< RAM0_CODE available address space  */
+#define RAM0_CODE_MEM_END          (0x10007FFFUL) /**< RAM0_CODE end address  */
+#define RAM0_CODE_MEM_BITS         (0x0000000FUL) /**< RAM0_CODE used bits  */
+#define CRYPTO0_MEM_BASE           (0x400F0000UL) /**< CRYPTO0 base address  */
+#define CRYPTO0_MEM_SIZE           (0x400UL)      /**< CRYPTO0 available address space  */
+#define CRYPTO0_MEM_END            (0x400F03FFUL) /**< CRYPTO0 end address  */
+#define CRYPTO0_MEM_BITS           (0x0000000AUL) /**< CRYPTO0 used bits  */
+#define PER_BITCLR_MEM_BASE        (0x44000000UL) /**< PER_BITCLR base address  */
+#define PER_BITCLR_MEM_SIZE        (0x50000UL)    /**< PER_BITCLR available address space  */
+#define PER_BITCLR_MEM_END         (0x4404FFFFUL) /**< PER_BITCLR end address  */
+#define PER_BITCLR_MEM_BITS        (0x00000013UL) /**< PER_BITCLR used bits  */
+#define PER1_BITCLR_MEM_BASE       (0x44050000UL) /**< PER1_BITCLR base address  */
+#define PER1_BITCLR_MEM_SIZE       (0xA0000UL)    /**< PER1_BITCLR available address space  */
+#define PER1_BITCLR_MEM_END        (0x440EFFFFUL) /**< PER1_BITCLR end address  */
+#define PER1_BITCLR_MEM_BITS       (0x00000014UL) /**< PER1_BITCLR used bits  */
+#define PER1_BITSET_MEM_BASE       (0x46050000UL) /**< PER1_BITSET base address  */
+#define PER1_BITSET_MEM_SIZE       (0xA0000UL)    /**< PER1_BITSET available address space  */
+#define PER1_BITSET_MEM_END        (0x460EFFFFUL) /**< PER1_BITSET end address  */
+#define PER1_BITSET_MEM_BITS       (0x00000014UL) /**< PER1_BITSET used bits  */
+#define CRYPTO0_BITSET_MEM_BASE    (0x460F0000UL) /**< CRYPTO0_BITSET base address  */
+#define CRYPTO0_BITSET_MEM_SIZE    (0x400UL)      /**< CRYPTO0_BITSET available address space  */
+#define CRYPTO0_BITSET_MEM_END     (0x460F03FFUL) /**< CRYPTO0_BITSET end address  */
+#define CRYPTO0_BITSET_MEM_BITS    (0x0000000AUL) /**< CRYPTO0_BITSET used bits  */
+#define CRYPTO0_BITCLR_MEM_BASE    (0x440F0000UL) /**< CRYPTO0_BITCLR base address  */
+#define CRYPTO0_BITCLR_MEM_SIZE    (0x400UL)      /**< CRYPTO0_BITCLR available address space  */
+#define CRYPTO0_BITCLR_MEM_END     (0x440F03FFUL) /**< CRYPTO0_BITCLR end address  */
+#define CRYPTO0_BITCLR_MEM_BITS    (0x0000000AUL) /**< CRYPTO0_BITCLR used bits  */
+#define PER_BITSET_MEM_BASE        (0x46000000UL) /**< PER_BITSET base address  */
+#define PER_BITSET_MEM_SIZE        (0x50000UL)    /**< PER_BITSET available address space  */
+#define PER_BITSET_MEM_END         (0x4604FFFFUL) /**< PER_BITSET end address  */
+#define PER_BITSET_MEM_BITS        (0x00000013UL) /**< PER_BITSET used bits  */
+#define PER_MEM_BASE               (0x40000000UL) /**< PER base address  */
+#define PER_MEM_SIZE               (0x50000UL)    /**< PER available address space  */
+#define PER_MEM_END                (0x4004FFFFUL) /**< PER end address  */
+#define PER_MEM_BITS               (0x00000013UL) /**< PER used bits  */
+#define PER1_MEM_BASE              (0x40050000UL) /**< PER1 base address  */
+#define PER1_MEM_SIZE              (0xA0000UL)    /**< PER1 available address space  */
+#define PER1_MEM_END               (0x400EFFFFUL) /**< PER1 end address  */
+#define PER1_MEM_BITS              (0x00000014UL) /**< PER1 used bits  */
+#define DEVICE_MEM_BASE            (0xF0040000UL) /**< DEVICE base address  */
+#define DEVICE_MEM_SIZE            (0x1000UL)     /**< DEVICE available address space  */
+#define DEVICE_MEM_END             (0xF0040FFFUL) /**< DEVICE end address  */
+#define DEVICE_MEM_BITS            (0x0000000CUL) /**< DEVICE used bits  */
+#define RAM0_MEM_BASE              (0x20000000UL) /**< RAM0 base address  */
+#define RAM0_MEM_SIZE              (0x8000UL)     /**< RAM0 available address space  */
+#define RAM0_MEM_END               (0x20007FFFUL) /**< RAM0 end address  */
+#define RAM0_MEM_BITS              (0x0000000FUL) /**< RAM0 used bits  */
 
 /** Single RAM space macros combining both RAM ports to match legacy, single-RAM-port chips */
-#define RAM_MEM_BASE               ((uint32_t) 0x20000000UL) /**< RAM base address  */
-#define RAM_MEM_SIZE               ((uint32_t) 0x8000UL)     /**< RAM available address space  */
-#define RAM_MEM_END                ((uint32_t) 0x20007FFFUL) /**< RAM end address  */
-#define RAM_MEM_BITS               ((uint32_t) 0x0000000FUL) /**< RAM used bits  */
+#define RAM_MEM_BASE               (0x20000000UL) /**< RAM base address  */
+#define RAM_MEM_SIZE               (0x8000UL)     /**< RAM available address space  */
+#define RAM_MEM_END                (0x20007FFFUL) /**< RAM end address  */
+#define RAM_MEM_BITS               (0x0000000FUL) /**< RAM used bits  */
 
 /** Flash and SRAM limits for EFM32TG11B540F64GM64 */
 #define FLASH_BASE                 (0x00000000UL) /**< Flash Base Address */
